keyboard-lock/navigator-cancelKeyLock.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

test(() => {
  assert_equals(navigator.cancelKeyLock(),
                undefined);
}, 'Keyboard Lock cancelKeyLock');

</script>
